Load browser meta only when browser opens
