+++ packages/dataclass-editor/src/detail/field/browser/typeBrowserApply.ts
@@ -1,7 +1,7 @@
 import type { BrowserNode } from '@axonivy/ui-components';
-import type { DataclassType, JavaType } from '../../protocol/types';
+import type { DataclassType, JavaType } from '../../../protocol/types';
 
-export const getApplyValue = (type: BrowserNode<DataclassType> | undefined, ivyTypes: JavaType[], typeAsList: boolean) => {
+export const typeBrowserApply = (type: BrowserNode<DataclassType> | undefined, ivyTypes: Array<JavaType>, typeAsList: boolean) => {
   if (type) {
     const fullQualifiedName = type.info + '.' + type.value;
     if (ivyTypes.some(ivyType => ivyType.fullQualifiedName === fullQualifiedName)) {
